requirements.txt
altair==5.5.0
annotated-types==0.7.0
asttokens==3.0.0
attrs==25.4.0
beautifulsoup4==4.14.2
blinker==1.9.0
cachetools==6.2.1
certifi==2025.10.5
chardet==3.0.4
charset-normalizer==3.4.4
click==8.1.8
colorama==0.4.6
decorator==5.2.1
docopt==0.6.2
elevenlabs==0.2.27
exceptiongroup==1.3.0
executing==2.2.1
gitdb==4.0.12
gitpython==3.1.45
googletrans==4.0.0rc1
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2025.1.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
ipython==8.18.1
jedi==0.19.2
jinja2==3.1.6
js2py==0.74
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
markupsafe==3.0.3
matplotlib-inline==0.1.7
narwhals==2.9.0
numpy==2.0.2
packaging==25.0
pandas==2.3.3
parso==0.8.5
pillow==11.3.0
pip==25.1.1
pipwin==0.5.2
prompt-toolkit==3.0.52
protobuf==6.33.0
pure-eval==0.2.3
pyarrow==21.0.0
pyaudio==0.2.14
pydantic==2.12.3
pydantic-core==2.41.4
pydeck==0.9.1
pygments==2.19.2
pyjsparser==2.7.1
pyprind==2.11.3
pysmartdl==1.3.4
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
referencing==0.36.2
requests==2.32.5
rfc3986==1.5.0
rpds-py==0.27.1
setuptools==80.3.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.8
speechrecognition==3.14.3
stack-data==0.6.3
streamlit==1.50.0
tenacity==9.1.2
toml==0.10.2
tornado==6.5.2
traitlets==5.14.3
typing-extensions==4.15.0
typing-inspection==0.4.2
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.5.0
watchdog==6.0.0
wcwidth==0.2.14
websockets==15.0.1
